handle React not behaving correctly wrt. key prop.
According to the docs, `key=` should cause dynamic children to refuse
to update to a sibling's value, preventing reuse of components across
dynamic children. Well, it's not doing that, and I'm not sure
why. This code works around that by placing initialization mechanisms
in update callbacks instead of mount callbacks.

This should be remedied, rather than worked around. The solution would
be cleaner.