fatter model; implement routing with app state; footer links follow routes
some of the model logic got moved into the model, appropriately. Like
getCompleted and getActive.

routing is checked at the app-level, and a `filter` is modified in the
app's state. This trickles down via props to all components that care
about what page we're on. Currently only footer filtering is
implemented. The rest should not take much code.